requirements.txt
amqp==2.3.2
astroid==2.1.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
idna==2.8
isort==4.3.4
kombu==4.2.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.2.2
python-dotenv==0.10.1
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.12.0
SQLAlchemy==1.2.15
typed-ast==1.1.1
urllib3==1.24.1
vine==1.1.4
wrapt==1.10.11
